(safe version)A function that frees a listint_t list..

+++ 0x13-more_singly_linked_lists/102-free_listint_safe.c
@@ -8,7 +8,7 @@
 
 size_t free_listint_safe(listint_t **h)
 {
-	unsigned int i, j, num = 0;
+	size_t i, j, num = 0;
 	listint_t buffer[1024];
 	listint_t *current = *h;
